+++ lib/services/dialog_service.dart
@@ -1,21 +1,40 @@
 import 'dart:async';
 
-class DialogService {
-  Function _showDialogListener;
-  Completer _dialogCompleter;
+class AlertRequest {
+  final String title;
+  final String description;
 
-  void registerDialogListener(Function showDialogListener) {
+  AlertRequest({this.title, this.description});
+}
+
+
+class AlertResponse {
+  final bool confirmed;
+  AlertResponse({this.confirmed});
+}
+
+
+class DialogService {
+  Function(AlertRequest) _showDialogListener;
+  Completer<AlertResponse> _dialogCompleter;
+
+  void registerDialogListener(Function(AlertRequest) showDialogListener) {
     _showDialogListener = showDialogListener;
   }
 
-  Future showDialog() {
-    _dialogCompleter = Completer();
-    _showDialogListener();
+  Future showDialog({
+    String title,
+    String description,
+  }) {
+    _dialogCompleter = Completer<AlertResponse>();
+    _showDialogListener(AlertRequest(title: title, description: description));
     return _dialogCompleter.future;
   }
-
-  void dialogComplete() {
-    _dialogCompleter.complete();
+  
+  void dialogComplete(AlertResponse response) {
+    if (_dialogCompleter != null) {
+      _dialogCompleter.complete(response);
+    }
     _dialogCompleter = null;
   }
 }